pyproject.toml
[project]
name = "ann-benchmark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.3",
    "psycopg2-binary>=2.9.10",
    "tqdm>=4.66.0",
]
